+++ src/device/tsp/Tsp/Script/SessionBaseExtensions/Methods_build.cs
@@ -104,8 +104,8 @@ public static partial class SessionBaseExtensionMethods
     }
 
     /// <summary>
-    /// A <see cref="Pith.SessionBase"/> extension method that loads the script from the deploy file and saves it to non-
-    /// volatile memory.
+    /// A <see cref="Pith.SessionBase"/> extension method that loads the script from the deploy file and saves it to
+    /// non-volatile memory.
     /// </summary>
     /// <remarks>   2025-04-15. </remarks>
     /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
@@ -132,6 +132,39 @@ public static partial class SessionBaseExtensionMethods
 
         SessionBaseExtensionMethods.TraceLastAction( $"Importing script from {scriptInfo.DeployFileFormat} '{deployFilePath}' file" );
         session.ImportScript( scriptInfo.Title, deployFilePath );
+
+        session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );
+
+        session.SaveScript( scriptInfo.Title );
+
+        session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );
+    }
+
+    /// <summary>
+    /// A <see cref="Pith.SessionBase"/> extension method that imports the scripts from a <see cref="StreamReader"/>
+    /// and saves it to non-volatile memory.
+    /// </summary>
+    /// <remarks>   2025-04-17. </remarks>
+    /// <exception cref="ArgumentNullException">        Thrown when one or more required arguments
+    ///                                                 are null. </exception>
+    /// <exception cref="InvalidOperationException">    Thrown when the requested operation is
+    ///                                                 invalid. </exception>
+    /// <param name="session">      The session. </param>
+    /// <param name="scriptInfo">   Information describing the script. </param>
+    /// <param name="reader">       The reader. </param>
+    public static void LoadSaveToNvm( this Pith.SessionBase session, ScriptInfoBase scriptInfo, TextReader reader )
+    {
+        if ( session is null ) throw new ArgumentNullException( nameof( session ) );
+        if ( scriptInfo is null ) throw new ArgumentNullException( nameof( scriptInfo ) );
+        if ( reader is null ) throw new ArgumentNullException( nameof( reader ) );
+
+        if ( !session.IsDeviceOpen ) throw new InvalidOperationException( $"{nameof( session )} is not open." );
+
+        // delete the script if it exists.
+        session.DeleteScript( scriptInfo.Title );
+
+        SessionBaseExtensionMethods.TraceLastAction( $"Importing script from {scriptInfo.DeployFileFormat} a '{reader.GetType()}'" );
+        session.ImportScript( reader, scriptInfo.Title );
 
         session.RunScript( scriptInfo.Title, scriptInfo.VersionGetterElement );